Introduce httpwg-macros, httpwg-gen
