+++ crates/fluke/tests/httpwg.rs
@@ -104,7 +104,7 @@ pub fn start_server() -> httpwg::Conn<TwoHalves<PipeWrite, PipeRead>> {
     httpwg::Conn::new(config, TwoHalves(client_write, client_read))
 }
 
-httpwg::gen_tests! {{
+httpwg_macros::gen_tests! {{
    crate::setup_tracing_and_error_reporting();
 
    fluke_buffet::start(async move {
