+++ crates/httpwg-gen/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "httpwg-gen"
+version = "0.1.0"
+edition = "2021"
+description = "Generates httpwg-macros from rustdoc's JSON output"
+
+[dependencies]
